+++ spec/lib/aixm/document_spec.rb
@@ -306,6 +306,7 @@ describe AIXM::Document do
             <txtLocalType>POLYGON</txtLocalType>
             <txtName>POLYGON AIRSPACE</txtName>
             <codeClass>C</codeClass>
+            <codeActivity>TFC-AD</codeActivity>
             <codeDistVerUpper>STD</codeDistVerUpper>
             <valDistVerUpper>65</valDistVerUpper>
             <uomDistVerUpper>FL</uomDistVerUpper>
@@ -363,6 +364,7 @@ describe AIXM::Document do
             <txtLocalType>CIRCLE</txtLocalType>
             <txtName>CIRCLE AIRSPACE</txtName>
             <codeClass>C</codeClass>
+            <codeActivity>TFC-AD</codeActivity>
             <codeDistVerUpper>STD</codeDistVerUpper>
             <valDistVerUpper>65</valDistVerUpper>
             <uomDistVerUpper>FL</uomDistVerUpper>
@@ -856,6 +858,7 @@ describe AIXM::Document do
             <txtLocalType>POLYGON</txtLocalType>
             <txtName>POLYGON AIRSPACE</txtName>
             <codeClass>C</codeClass>
+            <codeActivity>TFC-AD</codeActivity>
             <codeDistVerUpper>STD</codeDistVerUpper>
             <valDistVerUpper>65</valDistVerUpper>
             <uomDistVerUpper>FL</uomDistVerUpper>
@@ -914,6 +917,7 @@ describe AIXM::Document do
             <txtLocalType>CIRCLE</txtLocalType>
             <txtName>CIRCLE AIRSPACE</txtName>
             <codeClass>C</codeClass>
+            <codeActivity>TFC-AD</codeActivity>
             <codeDistVerUpper>STD</codeDistVerUpper>
             <valDistVerUpper>65</valDistVerUpper>
             <uomDistVerUpper>FL</uomDistVerUpper>