+++ spec/lib/aixm/feature/airspace_spec.rb
@@ -132,6 +132,7 @@ describe AIXM::Feature::Airspace do
             <txtLocalType>POLYGON</txtLocalType>
             <txtName>POLYGON AIRSPACE</txtName>
             <codeClass>C</codeClass>
+            <codeActivity>TFC-AD</codeActivity>
             <codeDistVerUpper>STD</codeDistVerUpper>
             <valDistVerUpper>65</valDistVerUpper>
             <uomDistVerUpper>FL</uomDistVerUpper>
@@ -195,6 +196,7 @@ describe AIXM::Feature::Airspace do
               <codeId>PA</codeId>
             </AseUid>
             <codeClass>C</codeClass>
+            <codeActivity>TFC-AD</codeActivity>
             <codeDistVerUpper>STD</codeDistVerUpper>
             <valDistVerUpper>65</valDistVerUpper>
             <uomDistVerUpper>FL</uomDistVerUpper>
@@ -307,6 +309,7 @@ describe AIXM::Feature::Airspace do
             </AseUid>
             <txtName>POLYGON AIRSPACE LAYER 1</txtName>
             <codeClass>C</codeClass>
+            <codeActivity>TFC-AD</codeActivity>
             <codeDistVerUpper>STD</codeDistVerUpper>
             <valDistVerUpper>65</valDistVerUpper>
             <uomDistVerUpper>FL</uomDistVerUpper>
@@ -344,6 +347,7 @@ describe AIXM::Feature::Airspace do
             </AseUid>
             <txtName>POLYGON AIRSPACE LAYER 2</txtName>
             <codeClass>C</codeClass>
+            <codeActivity>TFC-AD</codeActivity>
             <codeDistVerUpper>STD</codeDistVerUpper>
             <valDistVerUpper>65</valDistVerUpper>
             <uomDistVerUpper>FL</uomDistVerUpper>